pyproject.toml
[tool.poetry]
name = "cache-nix-action"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
